checkpoint-I/movie-service/src/main/java/com/example/movieservice/service/MovieService.java
package com.example.movieservice.service;

import com.example.movieservice.model.Movie;
import com.example.movieservice.repository.MovieRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class MovieService {

    @Autowired
    private MovieRepository movieRepository;

    public Movie save(Movie movie) {
        return movieRepository.save(movie);
    }

    public Movie findById(Long id) {
        return movieRepository.findById(id).orElse(null);
    }

    public void deleteById(Long id) {
        movieRepository.deleteById(id);
    }

    public Movie update(Movie movie) {
        return movieRepository.save(movie);
    }

    public List<Movie> findByGenre(String genre) {
        return movieRepository.findByGenre(genre);
    }

    public List<Movie> findAll() {
        return movieRepository.findAll();
    }
}
